requirements.txt
numpy==1.19.0
Pillow==7.2.0
